Update Libraries.txt
Added NRF24L01 library reference to mirror final release of the book.  This file is only a text file listing libraries and where they may be found on the Internet at the time of the book's publication, it is not the source code for the libraries which belongs to the author and may be released under varying licenses.